fix: compile errors — dose-based ColorCount, ActivityGrid wiring, imports
- ColorCount.count is now Double (dose sums); BarChart axis labels render
  integers when the value is integral
- SubstanceCompanionScreen: DailyCount carries daily dose (Double),
  dateDoseMap/maxDose wiring, accentColor param, restored
  isSystemInDarkTheme import
- StatsAnalysisScreen: duplicate Icons import removed

+++ app/src/main/java/com/isaakhanimann/journal/ui/tabs/stats/StatsViewModel.kt
@@ -277,7 +277,7 @@ data class StatsModel(
     val consumerName: String?
 )
 
-data class ColorCount(val color: AdaptiveColor, val count: Int)
+data class ColorCount(val color: AdaptiveColor, val count: Double)
 
 data class StatItem(
     val substanceName: String,
